extract common LifecycleHookHandler to core module
